+++ src/main/webapp/app/entities/programming-exercise/index.ts
@@ -14,3 +14,4 @@ export * from './status/programming-exercise-instructor-status.component';
 export * from './test-cases/programming-exercise-manage-test-cases.component';
 export * from './programming-exercise.route';
 export * from './services';
+export { TestCaseState } from 'app/entities/programming-exercise/instructions/programming-exercise-instruction.service';
